Reorganized some files and changed scrobbling from 4 hours -> 1 hour
