+++ API/Data/DataContext.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using API.Entities;
 using API.Entities.Enums;
 using API.Entities.Enums.UserPreferences;
+using API.Entities.History;
 using API.Entities.Interfaces;
 using API.Entities.Metadata;
 using API.Entities.Scrobble;
