+++ API/Data/ManualMigrations/ManualMigrateBlacklistTableToSeries.cs
@@ -2,6 +2,7 @@
 using System.Linq;
 using System.Threading.Tasks;
 using API.Entities;
+using API.Entities.History;
 using API.Entities.Metadata;
 using Kavita.Common.EnvironmentInfo;
 using Microsoft.EntityFrameworkCore;
