+++ API/Data/ManualMigrations/ManualMigrateEncodeSettings.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using API.Entities;
 using API.Entities.Enums;
+using API.Entities.History;
 using Flurl.Util;
 using Kavita.Common.EnvironmentInfo;
 using Microsoft.EntityFrameworkCore;
