+++ API/Data/ManualMigrations/ManualMigrateMixedSpecials.cs
@@ -3,6 +3,7 @@ using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using API.Entities;
+using API.Entities.History;
 using API.Extensions;
 using API.Helpers.Builders;
 using API.Services;
