+++ API/Data/ManualMigrations/MigrateChapterNumber.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Threading.Tasks;
 using API.Entities;
+using API.Entities.History;
 using API.Services.Tasks.Scanner.Parser;
 using Kavita.Common.EnvironmentInfo;
 using Microsoft.EntityFrameworkCore;
