+++ API/Data/ManualMigrations/MigrateManualHistory.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Threading.Tasks;
 using API.Entities;
+using API.Entities.History;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
