+++ API/Data/ManualMigrations/MigrateSmartFilterEncoding.cs
@@ -4,6 +4,7 @@ using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using API.DTOs.Filtering.v2;
 using API.Entities;
+using API.Entities.History;
 using API.Helpers;
 using Kavita.Common.EnvironmentInfo;
 using Microsoft.EntityFrameworkCore;
