+++ Kavita.Common/Kavita.Common.csproj
@@ -13,12 +13,12 @@
       <PackageReference Include="Cronos" Version="0.9.0" />
       <PackageReference Include="DotNet.Glob" Version="3.1.3" />
       <PackageReference Include="Flurl.Http" Version="4.0.2" />
-      <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="9.0.0" />
-      <PackageReference Include="Microsoft.Extensions.Hosting" Version="9.0.0" />
-      <PackageReference Include="SonarAnalyzer.CSharp" Version="10.4.0.108396">
+      <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="9.0.1" />
+      <PackageReference Include="Microsoft.Extensions.Hosting" Version="9.0.1" />
+      <PackageReference Include="SonarAnalyzer.CSharp" Version="10.5.0.109200">
         <PrivateAssets>all</PrivateAssets>
         <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
       </PackageReference>
-      <PackageReference Include="xunit.assert" Version="2.9.2" />
+      <PackageReference Include="xunit.assert" Version="2.9.3" />
     </ItemGroup>
 </Project>
