+++ UI/Web/src/app/_single-module/user-scrobble-history/user-scrobble-history.component.ts
@@ -101,7 +101,6 @@ export class UserScrobbleHistoryComponent implements OnInit {
     this.isLoading = true;
     this.cdRef.markForCheck();
 
-    console.log('load page with: ', {query, field, isDescending, page, pageSize})
     this.scrobblingService.getScrobbleEvents({query, field, isDescending}, page, pageSize)
       .pipe(take(1))
       .subscribe((result: PaginatedResult<ScrobbleEvent[]>) => {
